前端admin后台完成菜单管理页面，包含增删改查功能, 以及进行了响应式处理和手机端适配
后端对管理员登录接口进行了角色校验,优化了鉴权逻辑, jwt过滤器里会判断登录用户的角色是否是admin或viewer再把菜单角色权限信息放存入Security上下文中

mysql新增角色viewer和user, viewer是管理后台查看者, 只能查看不能操作

+++ sidifensen_blog_backend/src/test/java/com/sidifensen/JwtUtilsTest.java
@@ -41,7 +41,7 @@ public class JwtUtilsTest {
     public void testParseToken() {
         String token = jwtUtils.createToken(id,false);
 
-        Long id = jwtUtils.parseToken(token);
+        Integer id = jwtUtils.parseToken(token);
         System.out.println("userId: " + id);
 
     }
@@ -100,7 +100,7 @@ public class JwtUtilsTest {
         } catch (InterruptedException e) {
             e.printStackTrace();
         }
-        Long id = jwtUtils.parseToken(token);
+        Integer id = jwtUtils.parseToken(token);
         System.out.println(id);
 
     }
@@ -123,7 +123,7 @@ public class JwtUtilsTest {
         System.out.println(localDateTime);
         boolean verify = JWTUtil.verify(token, "sidifensen".getBytes());
         System.out.println("verify: " + verify);
-        Long id = jwtUtils.parseToken(token);
+        Integer id = jwtUtils.parseToken(token);
         System.out.println(id);
     }
